ChefAPI/requirements.txt
certifi==2019.6.16
chardet==3.0.4
cycler==0.10.0
future==0.17.1
idna==2.8
kiwisolver==1.1.0
matplotlib==3.1.1
numexpr==2.7.0
numpy==1.17.0
pandas==0.25.0
pandastable==0.12.1
PyChef==0.3.0
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
requests==2.22.0
six==1.12.0
urllib3==1.25.3
xlrd==1.2.0
